mm, vmscan: don't turn on cache_trim_mode at the highest scan priority
With cache_trim_mode on, reclaim logic doesn't bother reclaiming anon
pages.  However, it should be more careful to turn on the mode because
it's going to prevent anon pages from being reclaimed even if there are a
huge number of anon pages that are very cold and which should be
reclaimed.  Even worse, that can lead kswapd_failures to reach
MAX_RECLAIM_RETRIES and stopping kswapd until direct reclaim eventually
works to resume kswapd.

Link: https://lkml.kernel.org/r/20240208061825.36640-1-byungchul@sk.com
Signed-off-by: Byungchul Park <byungchul@sk.com>
Cc: Yu Zhao <yuzhao@google.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>